starts to split apart the file saver code
